test: ✅ Motion for Collpase and Menu (#26828)
* test: ✅ test menu and collapse motion

* chore: move openAnimation into collapse

* add test case

* rm extra openAnimation.tsx

* remove unused code
