+++ tests/utils.ts
@@ -1,4 +1,5 @@
 import MockDate from 'mockdate';
+import { act } from 'react-dom/test-utils';
 
 export function setMockDate(dateString = '2017-09-18T03:30:07.795') {
   MockDate.set(dateString);
@@ -10,4 +11,8 @@ export function resetMockDate() {
 
 const globalTimeout = global.setTimeout;
 
-export const sleep = (timeout = 0) => new Promise(resolve => globalTimeout(resolve, timeout));
+export const sleep = async (timeout = 0) => {
+  await act(async () => {
+    await new Promise(resolve => globalTimeout(resolve, timeout));
+  });
+};
